[refactor] SortableGridItem를 분리#174
- TTSTableGridView.tsx에 있던 SortableGridItem 컴포넌트를 다른 파일로 분리했습니다.
